+++ services/gateway/apis/http-version3/methods/transactionsDryRun.js
@@ -30,6 +30,7 @@ module.exports = {
 				optional: false,
 				type: 'object',
 				props: {
+					id: { type: 'string' },
 					module: { type: 'string' },
 					command: { type: 'string' },
 					fee: { type: 'string' },
